% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bc_fit_zeta}
\alias{bc_fit_zeta}
\title{BC outbreak data hierarchical intervention posterior}
\format{
An object of class \code{list} of length 1.
}
\usage{
bc_fit_zeta
}
\description{
Object created from `seir_model_fit` for BC data with hierarchical intervention.
For more details see the BC model fitting vignette:
\code{vignette("BC_outbreak_data_fitting", package = "cr0eso")}
}
\keyword{datasets}
