% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr0eso-package.R
\docType{package}
\name{cr0eso-package}
\alias{cr0eso-package}
\alias{cr0eso}
\title{The 'cr0eso' package.}
\description{
Bayesian hierarchical estimation of R0 from facility data
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
