% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tables.R
\name{create_pub_tables}
\alias{create_pub_tables}
\title{Create comparison table for multiple posteriors}
\usage{
create_pub_tables(location_labels = NULL, ...)
}
\arguments{
\item{...}{named list where each item is output of [covid_fit_seir]}
}
\value{
tibble of model results
}
\description{
Create a comparison table of posteriors generated from the same
data. For example you may wish to compare the R0 where an intervention is
assumed compared to another model where no intervention is assumed
}
\examples{
\dontrun{
create_pub_tables(mod)
create_pub_tables(intervention = mod,no_intervention= mod2)
}
}
