% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{hom_extract_posterior_draws}
\alias{hom_extract_posterior_draws}
\title{Extract model incidence and R0s from posterior}
\usage{
hom_extract_posterior_draws(posts)
}
\arguments{
\item{posts}{Object after calling extract of stan model object of hierarchical model}
}
\description{
Extract model incidence and R0s from posterior
}
\note{
Naming convention throughout is snake case with prefix "hom_" to denote Hierarcical Outbreak Model
}
\author{
Mike Irvine
}
