% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{hom_plot_counterfactual_by_location}
\alias{hom_plot_counterfactual_by_location}
\title{plot incidence and counterfactual scenario}
\usage{
hom_plot_counterfactual_by_location(
  fit,
  outbreak_cases = NULL,
  location_labels = NULL
)
}
\arguments{
\item{fit}{output of [seir_model_fit]}

\item{outbreak_cases}{matrix of outbreak case data}

\item{location_labels}{tibble with columns `site` and `location`. Will be used to
relabel numeric location to values in `site` column}
}
\value{
list containing:
 * plot - ggplot object
 * table - tibble object of results
 * violin_plot - ggplot object
}
\description{
on same plot show counterfactual and incidence posterior predictive distributions
across time and faceted by location. Shaded regions represent the 50% and 90% CrI, with median of
draws shown as lines.
}
\note{
Naming convention throughout is snake case with prefix "hom_" to denote Hierarcical Outbreak Model
}
\examples{
\dontrun{
 tmax <- 5
 pop_size <- 100
 dim(pop_size) <- c(1)
 example_incidence <- matrix(c(1,1,2,3,2),ncol=1)
 fit <- seir_model_fit(tmax,1,example_incidence,pop_size)
 result <- hom_plot_incidence_by_location(fit)
 # plot results
 show(result$plot)
 }
}
