% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{hom_plot_critical_times_by_location}
\alias{hom_plot_critical_times_by_location}
\title{Plot the critical time for each location}
\usage{
hom_plot_critical_times_by_location(
  posts,
  locations,
  num_cases,
  capacities,
  location_labels = NULL
)
}
\arguments{
\item{posts}{Object after calling extract of stan model object of hierarchical model}

\item{locations}{vector of location names}

\item{num_cases}{vector of number of cases for each location}

\item{capacities}{vector of sizes for each location}

\item{location_labels}{tibble with columns `site` and `location`. Will be used to
relabel numeric location to values in `site` column}
}
\description{
Critical time defined as time in days until the effective
 R number reduces to one or below. Plot also gives 95 credible interval and
 attack rate for each location.
}
\note{
Naming convention throughout is snake case with prefix "hom_" to denote Hierarcical Outbreak Model
}
