% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR_outbreak_model_fitting.R
\docType{data}
\name{prior_list}
\alias{prior_list}
\title{list of parameterisations for priors}
\format{
An object of class \code{list} of length 12.
}
\usage{
prior_list
}
\description{
list of parameterisations for priors
}
\keyword{datasets}
