% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR_outbreak_model_fitting.R
\name{seir_model_fit}
\alias{seir_model_fit}
\title{SEIR model fitting method to a multiple outbreak model}
\usage{
seir_model_fit(
  tmax,
  n_outbreaks,
  outbreak_cases,
  outbreak_sizes,
  intervention_switch = TRUE,
  multilevel_intervention = FALSE,
  priors = prior_list,
  iter = 600
)
}
\arguments{
\item{tmax}{Total number of time-points in observation}

\item{n_outbreaks}{Total number of outbreaks}

\item{outbreak_cases}{Numnber of daily reported cases by outbreak}

\item{outbreak_sizes}{The total size of each facility (initial number of suscepitble and exposed)}

\item{intervention_switch}{Describes whether interventions occur in data (default TRUE)}

\item{multilevel_intervention}{Describes whether intervention occurs}

\item{iter}{number of iterations of MCMC}

\item{prior_list}{List of priors. See [prior_list]}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Create an instance of the hierarchical SEIR stan model incorporating
various data elements and sample model.
}
\examples{
tmax <- 5
pop_size <- 100
dim(pop_size) <- c(1)
example_incidence <- matrix(c(1,1,2,3,2),ncol=1)
fit <- seir_model_fit(tmax,1,example_incidence,pop_size)
}
\author{
Mike Irvine
}
