% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BC_OSABC_facilitydata}
\alias{BC_OSABC_facilitydata}
\title{Additional covariate data on Long Term Health Care facility COVID-19 outbreaks in British Columbia}
\format{

}
\source{
Office of the Seniors Advocate, British Columbia \url{https://www.seniorsadvocatebc.ca/quickfacts/location} and British Columbia Centre for Disease Control
}
\usage{
BC_OSABC_facilitydata
}
\description{
A dataset containing additional covariate data for each of the 18 Long Term Health Care
facility COVID-19 outbreaks in British Columbia contained in dataset
'BC_LTHC_outbreaks_100Imputs'. The majority of this data was sourced from the Office of
the Seniors Advocate, British Columbia, Long-Term Care Facilities Quick Facts Directory.
https://www.seniorsadvocatebc.ca/quickfacts/location. All OSABC data is from the 2018/19
year. Identity of the initial case in each facility outbreak was obtained from the BCCDC.
}
\keyword{datasets}
