% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{hom_plot_incidence_by_location}
\alias{hom_plot_incidence_by_location}
\title{Plot incidence by location as extracted from hierarchical outbreak model.
If extracted_posts is NULL then posts object is used to first extract the
incidence draws from posterior (note this will take longer)}
\usage{
hom_plot_incidence_by_location(
  extracted_posts = NULL,
  posts = NULL,
  outbreak_cases = NULL,
  end_time = 60
)
}
\arguments{
\item{extracted_posts}{object returned by hom_extract_posterior_draws}

\item{posts}{Object after calling extract of stan model object of hierarchical model}

\item{outbreak_cases}{matrix of outbreak case data}

\item{end_time}{time in days to plot up until}
}
\value{
list containing:
 * plot - ggplot object
 * table - tibble object of results
}
\description{
Plot incidence by location as extracted from hierarchical outbreak model.
If extracted_posts is NULL then posts object is used to first extract the
incidence draws from posterior (note this will take longer)
}
\note{
Naming convention throughout is snake case with prefix "hom_" to denote Hierarcical Outbreak Model
}
\examples{
 mod # RSTAN fit object
 outbreak_cases # data of outbreaks in matrix form
 posts <- extract(mod) # extract posterior from model object
 extracted_posts <- hom_extract_posterior_draws(posts) # get object of incidence and r0
 result <- hom_plot_incidence_by_location(extracted_posts=extracted_posts,
                                          outbreak_cases=outbreak_cases)
 # plot results
 show(result$plot)
}
