% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR_outbreak_model_fitting.R
\name{seir_model_fit}
\alias{seir_model_fit}
\title{SEIR model fitting method to a multiple outbreak model}
\usage{
seir_model_fit(
  tmax,
  n_outbreaks,
  outbreak_cases,
  outbreak_sizes,
  intervention_switch = TRUE,
  multilevel_intervention = FALSE,
  iter = 600
)
}
\arguments{
\item{tmax}{- total number of time-points in observation}

\item{n_outbreaks}{- total number of outbreaks}

\item{outbreak_cases}{- Numnber of daily reported cases by outbreak}

\item{outbreak_sizes}{- The total size of each facility (initial number of suscepitble and exposed)}

\item{intervention_switch}{- Describes whether interventions occur in data (default TRUE)}

\item{multilevel_intervention}{- describes whether intervention occurs}

\item{iter}{number of iterations of MCMC}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Create an instance of the hierarchical SEIR stan model incorporating
various data elements and sample model.
}
\examples{
tmax <- 30
example_incidence <- c(1,1,2,3,2)
seir_model_fit(tmax,c(1),example_incidence,c(100))
}
\author{
Mike Irvine
}
